refactor(validate): HTML 워커를 추출하고 부모 포인터를 남긴다 (#214)
동작 변화 0. 기준선과 같은 112건이 그대로 통과한다.

PR #234 가 swatchFillCount 를 구조 인식 워커로 재작성하면서 "워커는 새 모듈로
빼지 않는다 — 실제 두 번째 소비자가 생길 때 순수 이동으로 빼낸다"고 적었다.
정부 식별자 규칙(#214)이 그 소비자다.

src/lib/html-walk.ts 로 옮긴 것: bodyOf · VOID_ELEMENTS · OPAQUE_ELEMENTS ·
WHITESPACE_ENTITIES · isBlankText · findTagEnd · parseAttrs, 그리고 메인 루프와
최종 unwind. 주석 안의 실측 근거와 보이지 않는 문자까지 보존해야 해서 손으로
옮겨 적지 않고 스크립트로 원문을 떠서 조립했다.

**추출하며 새로 생긴 것은 부모 포인터 하나다.** 옛 코드는 상속용으로 부모를
계산해 놓고 pop 할 때 노드를 버렸다. 요소가 닫힐 때 모든 판정이 끝나면 그걸로
충분하지만 #214 는 아니다 — 실제 마크업에서 캡션이 식별자보다 **뒤에** 오므로
순회가 끝나야 답을 알 수 있다. 부모를 유지하면 소비자가 노드를 모아뒀다가
나중에 판정할 수 있다.

**onText 는 필터하지 않고 원문을 넘긴다.** fill 계수는 isBlankText 로 걸러야
하지만 정부 규칙은 모든 텍스트에서 리터럴을 찾아야 한다. 필터를 워커에 넣으면
한 소비자를 조용히 섬기고 다른 소비자를 못 쓰게 만든다. 그래서 판정은
소비자 쪽에 남긴다.

swatchFillCount 는 소비자가 됐다 — markText 가 스택 대신 node.parent 를 타고,
record 는 onClose 로, theme 상속은 init 이 node.parent.data.theme 을 읽는다.
opaque 요소는 onText 를 onClose 보다 먼저 부른다(옛 push/markText/pop 대응) —
순서가 뒤집히면 자기 내용을 못 본다.

bodyOf 는 labelledTokenNames 도 쓰므로 새 모듈에서 export 하고 되import 했다.

Refs #214
